Deploy to `mvn.wesjd.net` (#391)
This PR switches the build and deployment to GH actions and GH pages. A
GH actions workflow builds the project in this repo and uploads the
files to `mvn.wesjd.net`. A GH action on `mvn.wesjd.net` deploys the
contents.

While we no longer have a dependency on CodeMC for build and
distribution, we still use their NMS repo. In the future, we may need to
have a more creative way to depend on the NMS code in the modules.

Closes #389